+++ frontend/src/app/calendar/page.tsx
@@ -1,4 +1,4 @@
-import Calendar from "@/components/Calendar";
+import Calendar from "@/components/calendar/Calendar";
 
 export default function MonthlyCalendar() {
     return (
